# Allow integer packing and unpacking of partial chunks