--- a/src/include/common/utils.h
+++ b/src/include/common/utils.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdint>
+#include <numeric>
 #include <vector>
 
 #include "common/assert.h"
@@ -13,10 +14,12 @@ namespace common {
 class BitmaskUtils {
 public:
     template<typename T>
-        requires std::integral<T> && std::is_unsigned_v<T>
+        requires std::integral<T>
     static T all1sMaskForLeastSignificantBits(uint32_t numBits) {
         KU_ASSERT(numBits <= 64);
-        return numBits == 64 ? std::numeric_limits<T>::max() : ((T)1 << numBits) - 1;
+        using U = numeric_utils::MakeUnSignedT<T>;
+        return std::bit_cast<T>(numBits == (sizeof(U) * 8) ? std::numeric_limits<U>::max() :
+                                                             static_cast<U>(((U)1 << numBits) - 1));
     }
 
     // constructs all 1s mask while avoiding overflow/underflow for int128