--- a/src/include/storage/compression/bitpacking_int128.h
+++ b/src/include/storage/compression/bitpacking_int128.h
@@ -4,7 +4,6 @@
 #pragma once
 
 #include "common/types/int128_t.h"
-#include "storage/compression/compression.h"
 
 namespace kuzu::storage {
 
@@ -14,15 +13,5 @@ struct Int128Packer {
     static void unpack(const uint32_t* __restrict in, common::int128_t* __restrict out,
         uint8_t width);
 };
-
-namespace bitpacking_utils {
-template<std::integral CompressedType, IntegerBitpackingType UncompressedType>
-void unpackSingle(const CompressedType* __restrict& in, UncompressedType* __restrict out,
-    uint16_t delta, uint16_t shiftRight);
-
-template<std::integral CompressedType, IntegerBitpackingType UncompressedType>
-void packSingle(const UncompressedType in, CompressedType* __restrict& out, uint16_t delta,
-    uint16_t shiftLeft, UncompressedType mask);
-}; // namespace bitpacking_utils
 
 } // namespace kuzu::storage