--- a/src/include/storage/compression/bitpacking_utils.h
+++ b/src/include/storage/compression/bitpacking_utils.h
@@ -13,15 +13,11 @@ struct BitpackingUtils {
         std::conditional_t<sizeof(UncompressedType) >= sizeof(uint32_t), uint32_t, uint8_t>;
     static constexpr size_t sizeOfCompressedTypeBits = sizeof(CompressedType) * 8;
 
-    static const uint8_t* unpackSingle(const uint8_t* __restrict src,
-        UncompressedType* __restrict dst, uint16_t bitWidth, size_t srcOffset);
+    static void unpackSingle(const uint8_t* __restrict src, UncompressedType* __restrict dst,
+        uint16_t bitWidth, size_t srcOffset);
 
-    static uint8_t* packSingle(const UncompressedType src, uint8_t* __restrict dstCursor,
+    static void packSingle(const UncompressedType src, uint8_t* __restrict dstCursor,
         uint16_t bitWidth, size_t dstOffset);
-
-    static const uint8_t* getInitialSrcCursor(const uint8_t* src, uint16_t bitWidth,
-        size_t srcOffset);
-    static uint8_t* getInitialSrcCursor(uint8_t* src, uint16_t bitWidth, size_t srcOffset);
 };
 
 } // namespace kuzu::storage