# Description

#2083

Reworked single-value bitpacking/unpacking for int128 to work for all integral types.

Used the above to allow packing of a number of values per page that isn't a multiple of 32 (by packing individual values at the end of each page).

# Contributor agreement

- [ ] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).