@royi-luo 2024-06-24
Some database sizes before/after the change (sizes in KB unless stated otherwise):

# After

## Tinysnb

```
du -ahk -d 1 ./workspace                                                                                                                                                                                                             py dev
84      ./workspace/n-0.hindex
44      ./workspace/n-1.hindex
4       ./workspace/rels.statistics
12      ./workspace/n-2.hindex.ovf
44      ./workspace/n-2.hindex
14140   ./workspace/data.kz
18708   ./workspace
```

## LDBC-sf01

du -ahk -d 1 ./workspace                                                                                                                                                                                                             py dev
7192    ./workspace/n-5.hindex
524     ./workspace/n-7.hindex
7568    ./workspace/n-0.hindex
2068    ./workspace/n-1.hindex
2060    ./workspace/n-3.hindex
2068    ./workspace/n-2.hindex
2068    ./workspace/n-6.hindex
2060    ./workspace/n-4.hindex
61504   ./workspace/data.kz
90012   ./workspace

## Very-large-serial

```
9148    ./workspace/n-0.hindex
960     ./workspace/data.kz
10160   ./workspace
```

## Custom Dataset, Single Column Int128 with bitwidth 118

```
du -ah -d 1 ./workspace                                                                                                                                                                                       py dev
4.4G    ./workspace/n-0.hindex
1.4G    ./workspace/data.kz
5.8G    ./workspace
```