@royi-luo 2024-06-25
> Can u also report some numbers on COPY? Just want to know if extra paddings will introduce some regression over COPY. The benchmark server is not able to show diff of COPY on ldbc100 right now.

```
create node table person (ID INT128, PRIMARY KEY (ID));
```

Quick performance benchmark on a dataset with 100M nodes:
| Action | Time On Branch (ms) | Time On Master (ms) |
| ---- | ---- | ---- |
| COPY (bitwidth 118) | 25987.65ms | 62947.98ms |
| MATCH (bitwidth 118) | 2554.96ms | 4029.66ms |
| COPY (bitwidth ~27) | 39939.65ms | 26193.03ms |
| MATCH (bitwidth ~27) | 5900.28ms | 2120.84ms |